Increased precision of values written to .wld file for WMS layers (bug 446)
git-svn-id: http://svn.osgeo.org/mapserver/trunk@3001 7532c77e-422f-0410-93f4-f0b67bdd69e2

+++ mapwmslayer.c
@@ -27,6 +27,9 @@
  * DEALINGS IN THE SOFTWARE.
  **********************************************************************
  * $Log$
+ * Revision 1.65  2004/04/17 06:33:24  dan
+ * Increased precision of values written to .wld file for WMS layers (bug 446)
+ *
  * Revision 1.64  2004/04/05 21:57:06  assefa
  * Fix bug in msWMSGetFeatureInfoURL not passing a wmsparams structure to the
  * msBuildWMSLayerURL.
@@ -1082,23 +1085,24 @@ int msDrawWMSLayerLow(int nLayerId, httpRequestObj *pasReqInfo,
             strcpy(wldfile+strlen(wldfile)-3, "wld");
         if (wldfile && (fp = fopen(wldfile, "wt")) != NULL)
         {
-            fprintf(fp, "%f\n", MS_CELLSIZE(pasReqInfo[iReq].bbox.minx,
-                                            pasReqInfo[iReq].bbox.maxx, 
-                                            map->width));
+            fprintf(fp, "%.12f\n", MS_CELLSIZE(pasReqInfo[iReq].bbox.minx,
+                                               pasReqInfo[iReq].bbox.maxx, 
+                                               map->width));
             fprintf(fp, "0\n");
             fprintf(fp, "0\n");
-            fprintf(fp, "%f\n", MS_CELLSIZE(pasReqInfo[iReq].bbox.maxy,
-                                            pasReqInfo[iReq].bbox.miny, 
-                                            map->height));
-            fprintf(fp, "%f\n", pasReqInfo[iReq].bbox.minx);
-            fprintf(fp, "%f\n", pasReqInfo[iReq].bbox.maxy);
+            fprintf(fp, "%.12f\n", MS_CELLSIZE(pasReqInfo[iReq].bbox.maxy,
+                                               pasReqInfo[iReq].bbox.miny, 
+                                               map->height));
+            fprintf(fp, "%.12f\n", pasReqInfo[iReq].bbox.minx);
+            fprintf(fp, "%.12f\n", pasReqInfo[iReq].bbox.maxy);
             fclose(fp);
 
             // GDAL should be called to reproject automatically.
             if (msDrawRasterLayerLow(map, lp, img) != 0)
                 status = MS_FAILURE;
 
-            unlink(wldfile);
+            if (!lp->debug)
+                unlink(wldfile);
         }
         else
         {
